requirements.txt
pyqt5>=5.15.0
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.20.0
pandas>=1.3.0